pyproject.toml
[project]
name = "sysneuro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.5.2",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "panda>=0.3.1",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
]
